+++ src/Gruffalo_Dimmer/inc/interrupt_HW.h
@@ -14,9 +14,9 @@
 // Initialises the GPIO Odd and Even interrupts.
 void initialiseInterrupts(void);
 // Registers a callback to a pin interrupt number. If the interrupt is already registered, it overwrites it.
-void setInterruptCallback(uint32_t pinNo, callbackPtr_t cbPtr);
+void setInterruptCallback(uint8_t pinNo, callbackPtr_t cbPtr);
 // Registers a callback with context to a pin interrupt number. If the interrupt is already registered, sets the next
 // free interrupt number.
-uint32_t setInterruptCallbackWCtx(uint32_t pinNo, callbackCtxPtr_t cbCtxPtr, void* ctxPtr);
+uint32_t setInterruptCallbackWCtx(uint8_t pinNo, callbackCtxPtr_t cbCtxPtr, void* ctxPtr);
 
 #endif // INTERRUPT_HW_H
